requirements.txt
asgiref==3.2.10
autopep8==1.5.7
certifi==2021.10.8
charset-normalizer==2.0.10
Django==3.1.2
django-cors-headers==3.5.0
django-filter==2.4.0
django-rest-passwordreset==1.2.1
djangorestframework==3.12.1
djangorestframework-simplejwt==4.7.2
idna==3.3
Markdown==3.3.2
psycopg2==2.9.1
pycodestyle==2.7.0
PyJWT==2.1.0
pytz==2020.1
requests==2.27.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.8
